+++ R/oauth-cache.R
@@ -152,36 +152,65 @@ token_match <- function(candidate, existing, package = "gargle") {
     return()
   }
 
-  candidate_email <- extract_email(candidate)
-  ## examples of possible values:
-  ## 'blah@example.org' an actual email
-  ## '*'                permission to use an email we find in the cache
-  ## ''                 no email and no instructions
-
-  ## if we have no instructions, we need user permission to consult the cache
-  if (empty_string(candidate_email) && !rlang::is_interactive()) {
-    return()
-  }
-
   m <- match2(candidate, existing)
   if (!is.na(m)) {
+    stopifnot(length(m) == 1)
     return(existing[[m]])
   }
+  # there is no full match
 
-  ## if email was specified and no full match, we're done
+  candidate_email <- extract_email(candidate)
+  # possible values    what they mean
+  # ------------------ ---------------------------------------------------------
+  # 'blah@example.org' user specified an email
+  # '*'                `email = TRUE`, i.e. permission to use *one* that we find
+  #                    (we still scold for multiple matches)
+  # ''                 user gave no email and no instructions
+
+  # if email was specified, we're done
   if (!empty_string(candidate_email) && candidate_email != "*") {
     return()
   }
-  ## possible scenarios:
-  ## candidate_email is '*'
-  ## candidate_email is '' and session is interactive
+  # candidate_email is either '*' or ''
 
-  ## match on the short hash
+  # match on the short hash
   m <- match2(mask_email(candidate), mask_email(existing))
+
+  # if no match on short hash, we're done
   if (anyNA(m)) {
     return()
   }
   existing <- existing[m]
+  # existing holds at least one short hash match
+
+  if (!rlang::is_interactive()) {
+    # proceed, but make sure user sees messaging about how to do
+    # non-interactive auth more properly
+    # https://github.com/r-lib/gargle/issues/92
+    withr::local_options(list(gargle_quiet = FALSE))
+    candidate_email <- "*"
+    if (length(existing) > 1) {
+      emails <- extract_email(existing)
+      emails <- glue("  * {emails}")
+      cat_line(glue(
+        "Suitable tokens found in the cache, associated with these emails:\n",
+        "{glue_collapse(emails, sep = '\n')}", "\n",
+        "The first will be used."
+      ))
+      existing <- existing[[1]]
+    }
+    msg <- c(
+      "Using an auto-discovered, cached token.\n",
+      "To suppress this message, modify your code or options to clearly ",
+      "consent to the use of a cached token.\n",
+      "See gargle's \"Non-interactive auth\" vignette for more details:\n",
+      "https://gargle.r-lib.org/articles/non-interactive-auth.html"
+    )
+    msg <- glue::glue_collapse(msg)
+    # morally, I'd like to throw a warning but current design of token_fetch()
+    # means warnings are caught
+    cat_line(msg)
+  }
 
   if (length(existing) == 1 && candidate_email == "*") {
     cat_line(glue(
@@ -189,14 +218,8 @@ token_match <- function(candidate, existing, package = "gargle") {
     ))
     return(existing)
   }
-  ## we need user to OK our discovery or pick from multiple emails
 
-  if (!rlang::is_interactive()) {
-    stop_need_user_interaction(
-      "Suitable cached tokens exist, but user confirmation is required."
-    )
-  }
-
+  # we need user to OK our discovery or pick from multiple emails
   withr::local_options(list(gargle_quiet = FALSE))
   emails <- extract_email(existing)
   cat_line(glue(